fixing warnings

+++ src/System/Filesystem.h
@@ -6,12 +6,12 @@
 namespace Gum {
 namespace Filesystem {
     extern void readFileContentsLineByLine(const File& file, const std::function<void(std::string line)>& func);
-    extern void readFileContentsInChucksBin(const File& file, const unsigned int& chunksize, const std::function<bool(const char* chunk, const unsigned int& readbytes)>& func);
+    extern void readFileContentsInChucksBin(const File& file, const size_t& chunksize, const std::function<bool(const char* chunk, const size_t& readbytes)>& func);
     extern std::string readFileContents(const File& file);
 
     extern void writeToFile(const File& file, const std::string& str);
     extern void appendToFile(const File& file, const std::string& str);
-    extern void appendToFileBin(const File& file, const char* data, const unsigned int& length);
+    extern void appendToFileBin(const File& file, const char* data, const size_t& length);
     
     extern void iterateThroughDirectory(const File& directory, const std::function<void(File entry)>& func);
     extern Filetype nativeTypeToFiletype(const std::string& path, const unsigned long& nativeData = 0);
